pyproject.toml
[project]
name = "frontend-chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dotenv>=0.9.9",
    "fastapi>=0.115.12",
    "google-generativeai>=0.8.5",
    "jinja2>=3.1.6",
    "motor>=3.7.1",
    "pymongo>=4.13.0",
    "pypdf2>=3.0.1",
    "python-multipart>=0.0.20",
    "uvicorn>=0.34.2",
]
